chore: adjusted renderer API types and constants

+++ include/redasm/surface/renderer.h
@@ -20,7 +20,7 @@ RD_API void rd_renderer_unkn(RDRenderer* self);
 RD_API void rd_renderer_loc(RDRenderer* self, RDAddress address,
                             unsigned int fill, RDNumberFlags flags);
 RD_API void rd_renderer_num(RDRenderer* self, i64 c, unsigned int base,
-                            usize fill, RDNumberFlags flags);
+                            unsigned int fill, RDNumberFlags flags);
 RD_API void rd_renderer_mnem(RDRenderer* self, const RDInstruction* instr,
                              RDThemeKind fg);
 RD_API void rd_renderer_reg(RDRenderer* self, int reg);
